fix(tests): eof: file rename
