+++ tests/prague/eip3540_eof_v1/spec.py
@@ -1,4 +1,5 @@
 """
 EOF V1 Constants used throughout all tests
 """
+
 EOF_FORK_NAME = "Prague"
