+++ tests/prague/eip3540_eof_v1/test_eof_example.py
@@ -9,7 +9,7 @@ from ethereum_test_tools import Opcodes as Op
 from ethereum_test_tools.eof.v1 import AutoSection, Container, Section
 from ethereum_test_tools.eof.v1.constants import NON_RETURNING_SECTION
 
-from .constants import EOF_FORK_NAME
+from .spec import EOF_FORK_NAME
 
 REFERENCE_SPEC_GIT_PATH = "EIPS/eip-3540.md"
 REFERENCE_SPEC_VERSION = "8dcb0a8c1c0102c87224308028632cc986a61183"
