+++ tests/prague/eip3540_eof_v1/test_execution_function.py
@@ -18,7 +18,7 @@ from ethereum_test_tools.eof.v1 import Container, Section
 from ethereum_test_tools.eof.v1.constants import MAX_CODE_SECTIONS, MAX_RETURN_STACK_HEIGHT
 from ethereum_test_tools.vm.opcode import Opcodes as Op
 
-from .constants import EOF_FORK_NAME
+from .spec import EOF_FORK_NAME
 
 # List all containers used within execution tests, since they will need to be
 # valid EOF V1 containers too
